requirement.txt
grpcio==1.51.1
ipfsapi==0.4.4
ipfshttpclient==0.8.0a2
iterators==0.2.0
jsons==1.6.3
numpy==1.24.1
protobuf==3.19.5
py-solc-x==1.1.1
pycrypto==2.6.1
pycryptodome==3.5.0
scikit-learn==1.2.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
web3==5.31.3
websockets==9.1
base64